Add variable `skip_s3_notification` to manage outside the module.
Since bucket notifications can only be managed once, if an implementer wants additional
notifications on the bucket, they must be managed outside this module. If you give this
variable as `true`, you *must* add a bucket notification to the lambda given in outputs
as `scan_lambda_function_arn`. See [this issue (#510) on the provider](https://github.com/hashicorp/terraform-provider-aws/issues/501#issuecomment-445106037)
for more details on the topic.

This patch does not include updated README.md because I see that terraform_docs is not
included in the pre-commit configuration, so I assume it is done by your pipeline.